requirements.txt
Django>=4.2
gunicorn
psycopg2-binary
python-dotenv
Pillow
whitenoise
django-crispy-forms
django-widget-tweaks
